+++ CHANGELOG.md
@@ -22,3 +22,32 @@
 #### v0.1.3
 
 - [x] package.json 최종 버젼 반영
+
+
+#### v0.1.4
+
+- [x] 부 라이브러리 (unique, node, assert) 들 호출을 함수 방식으로 바꾸다.
+- [x] unit test 를 jasmine 에서 mocha로 바꾸다.
+- [x] assert#createSpy() 추가.
+
+
+#### v0.1.5
+
+- [x] node#delivery :: then()에 여러개의 인수 전달.
+
+
+#### v0.1.6
+
+- [x] node#copy :: 파일 복사 구현.
+- [x] Promise function 이름에 _ 를 지우자.
+- [x] debug#createSpy()에 function을 집어 넣을 수 있고,
+  function을 대신 실행하고 결과 값을 저장해 둔다.
+- [x] debug#createSpy(): 결과 값 저장할때 unshift 순으로 저장한다.
+  고로 가장 최근 저장된 결과 값은 returns[0]일 것이다.
+
+
+#### v0.1.7
+
+- [x] gulpfile.js: atom.GitNote를 위한 release 추가.
+- [x] gulpfile.js: test를 위한 release 추가.
+- [x] debug.createSpy(): func 저장.